Get is kinda good to go working on post.
right now it is getting the whole table but already have a way to show it getting just one single ID. 


Post is like halfway done. I have a method to be able to get it. 
source: https://stackoverflow.com/questions/49821318/how-to-insert-data-into-mysql-table-using-node-and-express

Update: this might be the hardest one but we will see

Delete: should be easy enough

foreign/primary: I have no idea how to do that